+++ hybig/browse.py
@@ -691,12 +691,10 @@ def scale_paletted_1band(
         if len(colors) > 256:
             colors = colors[:-1]
     else:
-        # if there is no ndv, add one to the end of the colormap
-        dst_nodata = np.uint8(len(colors))
-        colors = [*colors, nodata_color]
-        # This check is done explicitly since some colormaps may be <256 elements
-        if len(colors) > 256:
-            colors = [*colors[:-2], nodata_color]
+        # Reserve the final entry before casting the index to uint8. A full
+        # 256-color input palette otherwise wraps the nodata index to zero.
+        colors = [*colors[:255], nodata_color]
+        dst_nodata = np.uint8(len(colors) - 1)
 
     norm = BoundaryNorm(levels, len(levels) - 1)